Added trait for InternalRewardedAgent
+ trait was made to update explicit state(subjective rewards).
Later for reinforcement learning purpose penalties will be needed to be
calculated and stored in agent's trajectory of game. From environment it
gets already calculated rewards, however for his own he takes his state
reward and somehow must store penalty - it could be done by special
update on state to add explicit penalty but it would lead to
complicating definition of agent's states (info set). Instead Agent is
responsible for tracking penalty and when performing trace commit adding
it to the trace